# Use ESP32-S3 vector extensions for LUT processing and diffing