--- a/src/board/epd_board_v7.c
+++ b/src/board/epd_board_v7.c
@@ -162,20 +162,21 @@ static void epd_board_deinit() {
 
   int tries = 0;
   while (!((pca9555_read_input(config_reg.port, 1) & 0xC0) == 0x80)) {
-    if (tries >= 500) {
+    if (tries >= 50) {
       ESP_LOGE("epdiy", "failed to shut down TPS65185!");
       break;
     }
     tries++;
     vTaskDelay(1);
-    printf("%X\n", pca9555_read_input(config_reg.port, 1));
   }
+
   // Not sure why we need this delay, but the TPS65185 seems to generate an interrupt after some time that needs to be cleared.
-  vTaskDelay(500);
+  vTaskDelay(50);
   pca9555_read_input(config_reg.port, 0);
   pca9555_read_input(config_reg.port, 1);
-  ESP_LOGI("epdiy", "going to sleep.");
   i2c_driver_delete(EPDIY_I2C_PORT);
+
+  gpio_uninstall_isr_service();
 }
 
 static void epd_board_set_ctrl(epd_ctrl_state_t *state, const epd_ctrl_state_t * const mask) {
@@ -213,8 +214,6 @@ static void epd_board_poweron(epd_ctrl_state_t *state) {
 
   while (!(pca9555_read_input(config_reg.port, 1) & CFG_PIN_PWRGOOD)) {
   }
-
-  printf("PG is up\n");
 
   ESP_ERROR_CHECK(tps_write_register(config_reg.port, TPS_REG_ENABLE, 0x3F));