Adds optimized versions for 1bpp difference LUT lookup, highlevel framebuffer diffing, and output line masking. 
Combined with 120MHz PSRAM (activate in the experimental options), we now get sub-second updates for a 1872x1404 display using epdiy V7.