@schuhumi 2024-02-13
@vroland I got it to work! Only my waveform has lots of ghosting now, I'll need to tweak it again.

Does the `epd_draw_base()` function use the `drawn_columns` after it returns? I had many problems with artifacts when modifying `drawn_columns` directly after calling `epd_draw_base()`, and could only remedy most of it by giving the function a copy of the drawn columns.

Also, are there any speed gains to be expected from using `drawn_columns` in small updates? At least by eye I couldn't observe any.