@vroland 2024-02-14
@schuhumi It should only be used when drawing is initiated, I can have a look again.

@martinberlin Indeed, looks like there is some alignment problem with some display resolutions, I'll look into it :)