@martinberlin 2024-02-22
New test with a 6" display:
epd_init(&epd_board_v7, &ED060XC3, EPD_LUT_64K);

```
I (2167) epdiy: highlevel diff area: x: 0, y: 0, w: 1024, h: 768
I (2167) epdiy: starting update, phases: 30
actual draw took 367ms.  // PSRAM at 80 Mhz
actual draw took 361ms.  // PSRAM at 120 Mhz
```

But independent of PSRAM I cannot see any better performance with this display. Or it's only when you have an existing framebuffer and you update with a new image where I can spot the time difference?
@schuhumi can you tell me how you did your test please?