@schuhumi 2024-03-24
@martinberlin Sorry for the late response, life had me quite busy..

So initially I did only try to get it to work, I did not do any benchmarks. But you got me curious, and I just had a closer look. I wrapped my draw_base function this way:

```c
uint32_t t1 = esp_timer_get_time() / 1000;
epd_draw_base(
    epd_full_screen(),
    fb,
    epd_full_screen(),
    MODE_DU | MODE_PACKING_1PPB_DIFFERENCE,
    temperature,
    NULL, // drawn_lines,
    NULL, // drawn columns (only when testing vector extensions)
    epd_get_display()->default_waveform
);
uint32_t t2 = esp_timer_get_time() / 1000;
printf("[<With/No> vector extensions] actual draw took %ldms.\n", t2 - t1);
``

For a fair comparison I made sure that in both cases:
- drawn_lines (and drawn_columns in the vector case) are set to NULL to always draw everything
- the same (4 sections long) waveform is used
- the same LcdEpdConfig_t.ckv_high_time of 90 used
- EpdDisplay_t.bus_speed set to 23 in both cases
- PSRAM speed was set to 120MHz in both cases.

Without vector extensions:
```
I (65260) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (65490) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (65830) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (66120) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (66860) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (67140) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (67950) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
I (68230) epdiy: starting update, phases: 4
[No Vector extensions] actual draw took 95ms.
```

With vector extensions:
```
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 93ms.
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 93ms.
[With vector extensions] actual draw took 92ms.
[With vector extensions] actual draw took 93ms.
```

So for me it's hardly any difference too... (this is on the ED133UT2 display)