@martinberlin 2024-03-29
Hello @vroland
glad that you are back!
Now that you fixed this buffer alignment is possible to test it with the 9.7" ?
Will try this weekend and submit more feedback editing this