@vroland 2024-03-29
Hi, yes I'm back and un-jetlagged again ;) The branch should now work with the 9.7" display. Afaik, displays with width % 16 != 0 never really worked before either. I think as a workaround we have to virtually increase resolution. I have no such display to test though, with your color display it just worked?

@schuhumi @martinberlin  Regarding the speed: With the LCD peripheral the output speed is fixed, and the computation has to keep up with whatever is set. To actually see a faster speed you have to increase the bus speed by calling `epd_set_lcd_pixel_clock_MHz(<something>);` or modifying `displays.c`. Then with the vector extensions you should be able to increase it further without seeing errors.