@martinberlin 2024-03-29
Hi @vroland great will try to increase the speed.

> Afaik, displays with width % 16 != 0 never really worked before either

About this I can confirm that those 2 models mentioned whose width in fact is not module 16, work perfectly with the **main** branch but they won’t work with this PR branch.
Even if you comment the assert(with%16 == 0) then it will work but make a ghost image displayed in the X access. Will add a photo later to show my point. So I’m quite sure there is an use case that is missing, since this happens with both of the color filter displays I can test.