@schuhumi 2024-04-02
Oh I see, now that's impressive! I was able to go from 23MHz to 30, at 32 I sometimes get line buffer underruns.  That decreases the time for epd_draw_base from 95ms to ~70ms!