@martinberlin 2024-04-21
> I added support for unaligned diffing and LUT lookup, i.e., all displays with width % 8 == 0 should now work. Can you test again?

This looks very good @vroland 
I will check it tomorrow with the display's I have including the Kaleido 6" that is .width = 1448 and let you know the results (Timings also with main vs vector branch)